fix: unite all division tag color usage to getDivisionTagColor
