+++ packages/web/src/features/executive/register-member/components/RegisterMemberTable.tsx
@@ -9,9 +9,9 @@ import { ApiReg019ResponseOk } from "@clubs/interface/api/registration/endpoint/
 
 import Table from "@sparcs-clubs/web/common/components/Table";
 import Tag from "@sparcs-clubs/web/common/components/Tag";
+import { getDivisionTagColor } from "@sparcs-clubs/web/constants/tableTagList";
 import {
   getTagColorFromClubType,
-  getTagColorFromDivision,
   getTagContentFromClubType,
 } from "@sparcs-clubs/web/types/clubdetail.types";
 
@@ -43,7 +43,7 @@ const columns = [
     id: "division.name",
     header: "분과",
     cell: info => {
-      const tagColor = getTagColorFromDivision(`${info.getValue()}`);
+      const tagColor = getDivisionTagColor(`${info.getValue()}`);
 
       return <Tag color={tagColor}>{info.getValue()}</Tag>;
     },
