+++ packages/web/src/features/my/components/MyMemberTable.tsx
@@ -9,10 +9,12 @@ import { ApiReg006ResponseOk } from "@clubs/interface/api/registration/endpoint/
 
 import Table from "@sparcs-clubs/web/common/components/Table";
 import Tag from "@sparcs-clubs/web/common/components/Tag";
-import { RegistrationStatusTagList } from "@sparcs-clubs/web/constants/tableTagList";
+import {
+  getDivisionTagColor,
+  RegistrationStatusTagList,
+} from "@sparcs-clubs/web/constants/tableTagList";
 import {
   getTagColorFromClubType,
-  getTagColorFromDivision,
   getTagContentFromClubType,
 } from "@sparcs-clubs/web/types/clubdetail.types";
 import { getTagDetail } from "@sparcs-clubs/web/utils/getTagDetail";
@@ -59,9 +61,7 @@ const columns = [
     id: "divisionName",
     header: "분과",
     cell: info => (
-      <Tag color={getTagColorFromDivision(info.getValue())}>
-        {info.getValue()}
-      </Tag>
+      <Tag color={getDivisionTagColor(info.getValue())}>{info.getValue()}</Tag>
     ),
     size: 10,
   }),
