+++ packages/web/src/types/clubdetail.types.ts
@@ -40,34 +40,4 @@ const getTagContentFromClubType = (
   return content;
 };
 
-const getTagColorFromDivision = (divisionName: string): TagColor => {
-  // TODO : getTagDetail 사용
-  switch (divisionName) {
-    case "생활문화":
-    case "사회":
-      return "GREEN";
-    case "연행예술":
-    case "종교":
-      return "BLUE";
-    case "전시창작":
-    case "구기체육":
-      return "ORANGE";
-    case "밴드음악":
-    case "생활체육":
-      return "PURPLE";
-    case "이공학술":
-    case "보컬음악":
-      return "PINK";
-    case "연주음악":
-    case "인문학술":
-      return "YELLOW";
-    default:
-      return "GREEN"; // 기본값 임의 지정
-  }
-};
-
-export {
-  getTagColorFromClubType,
-  getTagColorFromDivision,
-  getTagContentFromClubType,
-};
+export { getTagColorFromClubType, getTagContentFromClubType };
